Token tag rename and specialize constructors
Rename Token type_ to tag.
Token constructors for values now assign tag automatically depending on the argument type. In contrast, instruction token's integer value can only be passed as a plain int.
Updated constant keyword and symbol tables to the new constructors. Updated in place Token constructions too.

+++ language/Parser.h
@@ -9,7 +9,7 @@
 
 #define parserError(format, ...) printError("Parser: " format, __VA_ARGS__)
 
-typedef short int tok_size;
+typedef int tok_size;
 #define TOKEN_MAX 12000
 
 struct Parser
